Moved to Integer values

+++ app/Models/DailyMetricsModel.php
@@ -25,7 +25,7 @@ class DailyMetricsModel extends AbstractModel
             /** @var \Illuminate\Database\Schema\Blueprint $table */
             $table->increments('id');
             $table->unsignedSmallInteger('metric_id');
-            $table->float('value');
+            $table->integer('value');
             $table->unsignedSmallInteger('minute');
 
             $table->index('metric_id');
